% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vector_quality.R
\name{check_vector_quality}
\alias{check_vector_quality}
\title{Check for presence of NAs and non-numerics in a vector}
\usage{
check_vector_quality(x)
}
\arguments{
\item{x}{input data vector}
}
\value{
Boolean of whether the data is good to extract features on or not
}
\description{
Check for presence of NAs and non-numerics in a vector
}
