% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_multi_outputs}
\alias{demo_multi_outputs}
\title{Computed values for multi-feature classification results for use in vignette}
\format{
A list object
\describe{
  \item{FeatureSetResultsPlot}{ggplot comparing feature set classification accuracy}
  \item{TestStatistics}{data.frame of test statistics}
  \item{RawClassificationResults}{data.frame of raw classification accuracy}
}
}
\usage{
demo_multi_outputs
}
\description{
Format is:
}
\keyword{datasets}
