% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_outputs}
\alias{demo_outputs}
\title{Computed values for top features results for use in vignette}
\format{
A list object
\describe{
  \item{ResultsTable}{data.frame of results for top features}
  \item{FeatureFeatureCorrelationPlot}{ggplot heatmap of feature-feature correlation matrix}
  \item{ViolinPlots}{ggplot of distributions rendered as violins for top features}
}
}
\usage{
demo_outputs
}
\description{
Format is:
}
\keyword{datasets}
