% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature_list}
\alias{feature_list}
\title{All features available in theft in tidy format}
\format{
A tidy dataframe with 2 variables:
\describe{
  \item{feature_set}{Name of the set the feature is from}
  \item{feature}{Name of the feature}
}
}
\usage{
feature_list
}
\description{
The variables include:
}
\keyword{datasets}
