% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_features.R
\name{fft_features}
\alias{fft_features}
\title{Calculate FFT absolute value and angle coefficients for an input time-series vector}
\usage{
fft_features(y)
}
\arguments{
\item{y}{\code{numeric} vector of values}
}
\value{
\code{data.frame} of results
}
\description{
Computes the absolute value and angle of FFT coefficients 0-99, following the same convention as \code{tsfresh}'s \code{fft_coefficient} feature with \code{attr = "abs"} and \code{attr = "angle"}.
}
\references{
Christ, M., Braun, N., Neuffer, J., and Kempa-Liehr, A.W. (2018).
Time Series FeatuRe Extraction on basis of Scalable Hypothesis tests
(tsfresh -- A Python package). \emph{Neurocomputing}, \strong{307}, 72--77.
\doi{10.1016/j.neucom.2018.03.067}
}
\author{
Trent Henderson
}
