% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_good_features.R
\name{filter_good_features}
\alias{filter_good_features}
\title{Filter resample data sets according to good feature list}
\usage{
filter_good_features(data, x, good_features)
}
\arguments{
\item{data}{\code{list} of \code{"Train"} and \code{"Test"} data}

\item{x}{\code{integer} denoting the resample index to operate on}

\item{good_features}{\code{character} vector of good features to keep}
}
\value{
\code{list} of filtered train and test data
}
\description{
Filter resample data sets according to good feature list
}
\author{
Trent Henderson
}
