% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_good_features.R
\name{find_good_features}
\alias{find_good_features}
\title{Helper function to find features in both train and test set that are "good"}
\usage{
find_good_features(data, x)
}
\arguments{
\item{data}{\code{list} of \code{"Train"} and \code{"Test"} data}

\item{x}{\code{integer} denoting the resample index to operate on}
}
\value{
\code{character} vector of "good" feature names
}
\description{
Helper function to find features in both train and test set that are "good"
}
\author{
Trent Henderson
}
