% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gp.R
\name{fit_gp}
\alias{fit_gp}
\title{Fit a Gaussian process classification model}
\usage{
fit_gp(
  data,
  id_var = NULL,
  group_var = NULL,
  eps = 1e+06,
  iter = 3000,
  chains = 3,
  max_treedepth = 15
)
}
\arguments{
\item{data}{a tidy dataframe of feature results where each feature is a separate column}

\item{id_var}{a variable that uniquely identifies each observation}

\item{group_var}{a variable that denotes the categorical groups each observation relates to and is the target of prediction}

\item{eps}{Defaults to 1e6}

\item{iter}{the number of iterations to simulate per chain. Defaults to 3000}

\item{chains}{the number of Markov chains to fit. Defaults to 3}

\item{max_treedepth}{the maximum tree depth for the HMC algorithm. Defaults to 15}
}
\value{
an object of the class of model that was fit
}
\description{
Fit a Gaussian process classification model
}
\examples{
#' \dontrun{
}
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. http://mc-stan.org/.
}
\author{
Trent Henderson
}
