% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mixed_gam.R
\name{fit_mixed_gam}
\alias{fit_mixed_gam}
\title{Fit a mixed-effects GAM classification model}
\usage{
fit_mixed_gam(data)
}
\arguments{
\item{data}{a tidy dataframe of feature results where each feature is a separate column}

\item{id_var}{a variable that uniquely identifies each observation}

\item{group_var}{a variable that denotes the categorical groups each observation relates to and is the target of prediction}
}
\value{
an object of the class of model that was fit
}
\description{
Fit a mixed-effects GAM classification model
}
\examples{
#' \dontrun{
}
}
\references{
Wood, S.N. (2011) Fast stable restricted maximum likelihood and marginal likelihood estimation of semiparametric generalized linear models. Journal of the Royal Statistical Society (B) 73(1):3-36
}
\author{
Trent Henderson
}
