% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit classification model and compute key metrics}
\usage{
fit_models(data, iter_data, row_id, is_null_run = FALSE, classifier)
}
\arguments{
\item{data}{\code{list} containing train and test sets}

\item{iter_data}{\code{data.frame} containing the values to iterate over for seed and either feature name or set name}

\item{row_id}{\code{integer} denoting the row ID for \code{iter_data} to filter to}

\item{is_null_run}{\code{Boolean} whether the calculation is for a null model. Defaults to \code{FALSE}}

\item{classifier}{\code{function} specifying the classifier to fit. Should be a function with 2 arguments: \code{formula} and \code{data}. Please note that \code{tsfeature_classifier} z-scores data prior to modelling using the train set's information so disabling default scaling if your function uses it is recommended.}
}
\value{
\code{data.frame} of classification results
}
\description{
Fit classification model and compute key metrics
}
\author{
Trent Henderson
}
