% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_multivariate_classifier.R
\name{fit_multivariate_classifier}
\alias{fit_multivariate_classifier}
\title{Fit a classifier to feature matrix using all features or all features by set}
\usage{
fit_multivariate_classifier(
  data,
  id_var = "id",
  group_var = "group",
  by_set = FALSE,
  num_splits = 10,
  test_method = c("linear svm", "rbf svm")
)
}
\arguments{
\item{data}{the dataframe containing the raw feature matrix}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to "id"}

\item{group_var}{a string specifying the grouping variable that the data aggregates to. Defaults to "group"}

\item{by_set}{Boolean specifying whether to compute classifiers for each feature set. Defaults to FALSE}

\item{num_splits}{an integer specifying the number of 75/25 train-test splits to perform. Defaults to 10}

\item{test_method}{the algorithm to use for quantifying class separation}
}
\value{
an object of class list containing summaries of the classification models
}
\description{
Fit a classifier to feature matrix using all features or all features by set
}
\examples{
\dontrun{
featMat <- calculate_features(data = simData, 
  id_var = "id", 
  time_var = "timepoint", 
  values_var = "values", 
  group_var = "process", 
  feature_set = "catch22")
  
fit_multivariate_classifier(featMat,
  id_var = "id",
  group_var = "group",
  by_set = FALSE,
  num_splits = 10,
  test_method = "linear svm") 
}

}
\author{
Trent Henderson
}
