% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nn.R
\name{fit_nn}
\alias{fit_nn}
\title{Fit a neural network classifier}
\usage{
fit_nn(data, stepmax = 1e+06)
}
\arguments{
\item{data}{a tidy dataframe of feature results where each feature is a separate column}

\item{id_var}{a variable that uniquely identifies each observation}

\item{group_var}{a variable that denotes the categorical groups each observation relates to and is the target of prediction}
}
\value{
an object of the class of model that was fit
}
\description{
Fit a neural network classifier
}
\examples{
#' \dontrun{
}
}
\references{
Stefan Fritsch, Frauke Guenther and Marvin N. Wright (2019). neuralnet: Training of Neural Networks. R package version 1.44.2. https://CRAN.R-project.org/package=neuralnet
}
\author{
Trent Henderson
}
