% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rf.R
\name{fit_rf}
\alias{fit_rf}
\title{Fit a random forest classifier}
\usage{
fit_rf(data, ntree = 100, importance = TRUE)
}
\arguments{
\item{data}{a tidy dataframe of feature results where each feature is a separate column}

\item{id_var}{a variable that uniquely identifies each observation}

\item{group_var}{a variable that denotes the categorical groups each observation relates to and is the target of prediction}
}
\value{
an object of the class of model that was fit
}
\description{
Fit a random forest classifier
}
\examples{
#' \dontrun{
}
}
\references{
A. Liaw and M. Wiener (2002). Classification and Regression by randomForest. R News 2(3), 18--22.
}
\author{
Trent Henderson
}
