% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svm.R
\name{fit_svm}
\alias{fit_svm}
\title{Fit a support vector machine (SVM) classification model}
\usage{
fit_svm(data, kernel = "radial")
}
\arguments{
\item{data}{a tidy dataframe of feature results where each feature is a separate column}

\item{id_var}{a variable that uniquely identifies each observation}

\item{group_var}{a variable that denotes the categorical groups each observation relates to and is the target of prediction}
}
\value{
an object of the class of model that was fit
}
\description{
Fit a support vector machine (SVM) classification model
}
\examples{
#' \dontrun{
}
}
\references{
David Meyer, Evgenia Dimitriadou, Kurt Hornik, Andreas Weingessel and Friedrich Leisch (2019). e1071: Misc Functions of the Department of Statistics, Probability Theory Group (Formerly: E1071), TU Wien. R package version 1.7-3. https://CRAN.R-project.org/package=e1071
}
\author{
Trent Henderson
}
