% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rescale_vals.R
\name{get_rescale_vals}
\alias{get_rescale_vals}
\title{Calculate central tendency and spread values for all numeric columns in a dataset}
\usage{
get_rescale_vals(data)
}
\arguments{
\item{data}{\code{data.frame} containing data to normalise}
}
\value{
\code{list} of central tendency and spread values
}
\description{
Calculate central tendency and spread values for all numeric columns in a dataset
}
\author{
Trent Henderson
}
