% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_theft.R
\name{init_theft}
\alias{init_theft}
\title{Communicate to R the correct Python version containing the relevant libraries for calculating features}
\usage{
init_theft(path_to_python)
}
\arguments{
\item{path_to_python}{a string specifying the filepath to the version of Python containing the relevant libraries for calculating features}
}
\value{
no return value; called for side effects
}
\description{
Communicate to R the correct Python version containing the relevant libraries for calculating features
}
\author{
Trent Henderson
}
