% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_theft.R
\name{init_theft_kats}
\alias{init_theft_kats}
\title{Communicate to R the Python virtual environment containing kats only}
\usage{
init_theft_kats(venv)
}
\arguments{
\item{venv}{\code{character} specifying the name of the to the Python virtual environment where \code{"kats"} is installed}
}
\value{
no return value; called for side effects
}
\description{
Communicate to R the Python virtual environment containing kats only
}
\examples{
\dontrun{
install_python_pkgs("theft-test")
init_theft_kats("theft-test")
}

}
\author{
Trent Henderson
}
