% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_calcs.R
\name{kurtosis}
\alias{kurtosis}
\title{Calculate a kurtosis of a vector}
\usage{
kurtosis(y)
}
\arguments{
\item{y}{\code{numeric} vector of values}
}
\value{
\code{numeric} scalar of kurtosis
}
\description{
Calculate a kurtosis of a vector
}
\author{
Trent Henderson
}
