% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_interval.R
\name{make_title}
\alias{make_title}
\title{Helper function for converting to title case}
\usage{
make_title(x)
}
\arguments{
\item{x}{\code{character} vector}
}
\value{
\code{character} vector
}
\description{
Helper function for converting to title case
}
\author{
Trent Henderson
}
