% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalers.R
\name{mean_scaler}
\alias{mean_scaler}
\title{This function rescales a vector of numerical values by subtracting the mean}
\usage{
mean_scaler(x)
}
\arguments{
\item{x}{a numeric vector, preferably of feature values computed by other {theft} package functions}
}
\description{
This function rescales a vector of numerical values by subtracting the mean
}
\examples{
x <- 1 + 0.5 * 1:1000 + arima.sim(list(ma = 0.5), n = 1000)
outs <- mean_scaler(x)

}
\author{
Trent Henderson
}
