% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalers.R
\name{minmax_scaler}
\alias{minmax_scaler}
\title{This function rescales a vector of numerical values into the unit interval [0,1]
\deqn{z_i = \frac{x_i - \min(\mathbf{x})}{\max(\mathbf{x}) - \min(\mathbf{x})}}}
\usage{
minmax_scaler(x)
}
\arguments{
\item{x}{a numeric vector, preferably of feature values computed by other \code{theft} package functions}
}
\value{
\code{numeric} vector
}
\description{
This function rescales a vector of numerical values into the unit interval [0,1]
\deqn{z_i = \frac{x_i - \min(\mathbf{x})}{\max(\mathbf{x}) - \min(\mathbf{x})}}
}
\author{
Trent Henderson
}
