% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{moments}
\title{Calculate a basic set of the four moments of the distribution}
\usage{
moments(y)
}
\arguments{
\item{y}{\code{numeric} vector of values}
}
\value{
\code{data.frame} of results
}
\description{
Calculate a basic set of the four moments of the distribution
}
\author{
Trent Henderson
}
