% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_feature_frame.R
\name{normalise_feature_frame}
\alias{normalise_feature_frame}
\title{Scale each feature vector into a user-specified range for visualisation and modelling}
\usage{
normalise_feature_frame(
  data,
  names_var = NULL,
  values_var = NULL,
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax", "MeanSubtract")
)
}
\arguments{
\item{data}{a dataframe with at least 2 columns: names variable (feature names) and value variable}

\item{names_var}{a string denoting the name of the variable/column that holds the feature names}

\item{values_var}{a string denoting the name of the variable/column that holds the numerical feature values}

\item{method}{a rescaling/normalising method to apply. Defaults to 'RobustSigmoid'}
}
\value{
a dataframe with the value column rescaled into the specified range
}
\description{
Scale each feature vector into a user-specified range for visualisation and modelling
}
\examples{
\dontrun{
library(dplyr)
library(tsibbledata)

d <- tsibbledata::aus_retail \%>\%
  rename(Series_ID = 3)

feature_matrix <- calculate_features(data = d, 
  id_var = "Series_ID", 
  time_var = "Month", 
  values_var = "Turnover", 
  group_var = "State",
  feature_set = "catch22")
  
normed <- normalise_feature_frame(feature_matrix, names_var = "names", 
  values_var = "values", method = "RobustSigmoid")
}

}
\author{
Trent Henderson
}
