% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_feature_vector.R
\name{normalise_feature_vector}
\alias{normalise_feature_vector}
\title{Scale each value into a user-specified range for visualisation and analysis}
\usage{
normalise_feature_vector(
  x,
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax", "MeanSubtract")
)
}
\arguments{
\item{x}{a vector of scalar values}

\item{method}{a rescaling/normalising method to apply. Defaults to 'RobustSigmoid'}
}
\value{
a vector of scalar values normalised into the selected range
}
\description{
Scale each value into a user-specified range for visualisation and analysis
}
\examples{
data <- 1 + 0.5 * 1:1000 + arima.sim(list(ma = 0.5), n = 1000)
outs <- normalise_feature_vector(data, method = "RobustSigmoid")

}
\author{
Trent Henderson
}
