% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_feature_vector.R
\name{normalize_feature_vector}
\alias{normalize_feature_vector}
\title{Scale each value into a user-specified range for visualisation and analysis}
\usage{
normalize_feature_vector(
  x,
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax", "MeanSubtract")
)
}
\arguments{
\item{x}{a vector of scalar values}

\item{method}{a rescaling/normalising method to apply. Defaults to 'RobustSigmoid'}
}
\value{
a vector of scalar values normalised into the selected range
}
\description{
Scale each value into a user-specified range for visualisation and analysis
}
\examples{
library(dplyr)
library(tsibbledata)

d <- tsibbledata::aus_retail \%>\%
  rename(Series_ID = 3)

feature_matrix <- calculate_features(data = d, 
  id_var = "Series_ID", 
  time_var = "Month", 
  values_var = "Turnover", 
  group_var = "State",
  feature_set = "catch22")
  
x <- feature_matrix \%>\%
  filter(names == "DN_HistogramMode_5") \%>\%
  pull(values)
  
xnormed <- normalise_feature_vector(x, method = "MinMax")

}
\author{
Trent Henderson
}
