% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.low_dimension}
\alias{plot.low_dimension}
\title{Produce a plot for a low_dimension object}
\usage{
\method{plot}{low_dimension}(x, show_covariance = TRUE, ...)
}
\arguments{
\item{x}{the \code{low_dimension} object containing the dimensionality reduction projection calculated by \code{reduce_dims}}

\item{show_covariance}{\code{Boolean} of whether covariance ellipses should be shown on the plot. Defaults to \code{TRUE}}

\item{...}{Arguments to be passed to methods}
}
\value{
object of class \code{ggplot} that contains the graphic
}
\description{
Produce a plot for a low_dimension object
}
\author{
Trent Henderson
}
