% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation_matrix.R
\name{plot_correlation_matrix}
\alias{plot_correlation_matrix}
\title{Produce a correlation matrix plot showing pairwise correlations of time series with automatic hierarchical clustering}
\usage{
plot_correlation_matrix(
  data,
  is_normalised = FALSE,
  id_var = "id",
  values_var = "values",
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax"),
  cor_method = c("pearson", "spearman"),
  interactive = FALSE
)
}
\arguments{
\item{data}{a dataframewith at least 2 columns for 'id' and 'values' variables}

\item{is_normalised}{a Boolean as to whether the input feature values have already been scaled. Defaults to FALSE}

\item{id_var}{a string specifying the ID variable to compute pairwise correlations between. Defaults to "id"}

\item{values_var}{a string denoting the name of the variable/column that holds the numerical feature values. Defaults to "values"}

\item{method}{a rescaling/normalising method to apply. Defaults to 'RobustSigmoid'}

\item{cor_method}{the correlation method to use. Defaults to 'pearson'}

\item{interactive}{a Boolean as to whether to plot an interactive plotly graphic. Defaults to FALSE}
}
\value{
an object of class ggplot that contains the correlation matrix graphic
}
\description{
Produce a correlation matrix plot showing pairwise correlations of time series with automatic hierarchical clustering
}
\examples{
\dontrun{
featMat <- calculate_features(data = simData, 
  id_var = "id", 
  time_var = "timepoint", 
  values_var = "values", 
  group_var = "process", 
  feature_set = "catch22")
  
plot_correlation_matrix(data = featMat, 
  is_normalised = FALSE, 
  id_var = "id", 
  values_var = "values",
  method = "RobustSigmoid",
  cor_method = "person",
  interactive = FALSE)
}

}
\author{
Trent Henderson
}
