% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_discrimination.R
\name{plot_feature_discrimination}
\alias{plot_feature_discrimination}
\title{Produce a matrix of violin plots for each feature to show class discrimination}
\usage{
plot_feature_discrimination(
  data,
  id_var = "id",
  group_var = "group",
  features = "all",
  normalise = FALSE,
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax")
)
}
\arguments{
\item{data}{the dataframe containing the raw feature matrix}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to "id"}

\item{group_var}{a string specifying the grouping variable that the data aggregates to. Defaults to "group"}

\item{features}{a string or vector of strings specifying which features to filter by. Defaults to "all"}

\item{normalise}{a Boolean of whether to normalise features before plotting. Defaults to FALSE}

\item{method}{a rescaling/normalising method to apply if normalise = TRUE. Defaults to 'RobustSigmoid'}
}
\value{
an object of class ggplot containing the graphics
}
\description{
Produce a matrix of violin plots for each feature to show class discrimination
}
\examples{
\dontrun{
featMat <- calculate_features(data = simData, 
  id_var = "id", 
  time_var = "timepoint", 
  values_var = "values", 
  group_var = "process", 
  feature_set = "catch22")
  
plot_feature_discrimination(featMat,
  id_var = "id",
  group_var = "group",
  features = "all") 
}

}
\author{
Trent Henderson
}
