% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_matrix.R
\name{plot_feature_matrix}
\alias{plot_feature_matrix}
\title{Produce a heatmap matrix of the calculated feature value vectors and each unique time series with automatic hierarchical clustering.}
\usage{
plot_feature_matrix(
  data,
  is_normalised = FALSE,
  id_var = "id",
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax"),
  interactive = FALSE
)
}
\arguments{
\item{data}{a dataframe with at least 2 columns called 'names' and 'values'}

\item{is_normalised}{a Boolean as to whether the input feature values have already been scaled. Defaults to FALSE}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to "id"}

\item{method}{a rescaling/normalising method to apply. Defaults to 'RobustSigmoid'}

\item{interactive}{a Boolean as to whether to plot an interactive plotly graphic. Defaults to FALSE}
}
\value{
an object of class ggplot that contains the heatmap graphic
}
\description{
Produce a heatmap matrix of the calculated feature value vectors and each unique time series with automatic hierarchical clustering.
}
\examples{
\dontrun{
featMat <- calculate_features(data = simData, 
  id_var = "id", 
  time_var = "timepoint", 
  values_var = "values", 
  group_var = "process", 
  feature_set = "catch22")

plot_feature_matrix(featMat, 
  is_normalised = FALSE, 
  id_var = "id", 
  method = "MinMax",
  interactive = FALSE)
}

}
\author{
Trent Henderson
}
