% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_low_dimension.R
\name{plot_low_dimension}
\alias{plot_low_dimension}
\title{Produce a principal components analysis (PCA) on normalised feature values and render a bivariate plot to visualise it}
\usage{
plot_low_dimension(
  data,
  is_normalised = FALSE,
  id_var = NULL,
  group_var = NULL,
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax", "MeanSubtract"),
  plot = TRUE
)
}
\arguments{
\item{data}{a dataframe with at least 2 columns called 'names' and 'values'}

\item{is_normalised}{a Boolean as to whether the input feature values have already been scaled. Defaults to FALSE}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to NULL}

\item{group_var}{a string specifying the grouping variable that the data aggregates to. Defaults to NULL}

\item{method}{a rescaling/normalising method to apply. Defaults to 'RobustSigmoid'}

\item{plot}{a Boolean as to whether a bivariate plot should be returned or the calculation dataframe. Defaults to TRUE}
}
\value{
if plot = TRUE, returns an object of class ggplot, if plot = FALSE returns an object of class dataframe with PCA results
}
\description{
Produce a principal components analysis (PCA) on normalised feature values and render a bivariate plot to visualise it
}
\examples{
\dontrun{
data1 <- 1 + 0.5 * 1:1000 + arima.sim(list(ma = 0.5), n = 1000)
data2 <- rnorm(1000, mean = 0, sd = 1)
outs1 <- catch22_all(data1)
outs1['group'] <- 'Group 1'
outs2 <- catch22_all(data2)
outs2['group'] <- 'Group 2'
outs <- rbind(outs1, outs2)
plot_low_dimension(outs, is_normalised = FALSE, id_var = "group", group = NULL, method = "RobustSigmoid", plot = TRUE)
}

}
\author{
Trent Henderson
}
