% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quality_matrix.R
\name{plot_quality_matrix}
\alias{plot_quality_matrix}
\title{Produce a matrix visualisation of data types computed by feature calculation function.}
\usage{
plot_quality_matrix(data, ignore_good_features = FALSE)
}
\arguments{
\item{data}{a dataframe with at least 2 columns called \code{"names"} and \code{"values"}}

\item{ignore_good_features}{Boolean whether to remove "good" features (i.e., successful numeric values) from the plot. Defaults to \code{FALSE}}
}
\value{
an object of class \code{ggplot}
}
\description{
Produce a matrix visualisation of data types computed by feature calculation function.
}
\examples{
featMat <- calculate_features(data = simData, 
  id_var = "id", 
  time_var = "timepoint", 
  values_var = "values", 
  group_var = "process", 
  feature_set = "catch22",
  seed = 123)

plot_quality_matrix(data = featMat,
  ignore_good_features = FALSE)

}
\author{
Trent Henderson
}
