% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_bayes_outputs.R
\name{print_bayes_outputs}
\alias{print_bayes_outputs}
\title{Produce a PDF report of all the model fit and diagnostic plots for a Bayesian classification program}
\usage{
print_bayes_outputs(model)
}
\arguments{
\item{model}{a model output from one of the Bayes function arguments contained in {sawlog}}
}
\value{
a PDF document saved in the current work directory
}
\description{
Produce a PDF report of all the model fit and diagnostic plots for a Bayesian classification program
}
\examples{
\dontrun{
print_bayes_outputs(model)
}

}
\author{
Trent Henderson
}
