% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hctsa_file.R
\name{process_hctsa_file}
\alias{process_hctsa_file}
\title{Load in hctsa formatted MATLAB files of time series data into a tidy format ready for feature extraction}
\usage{
process_hctsa_file(data)
}
\arguments{
\item{data}{\code{string} specifying the filepath to the \code{MATLAB} file to parse}
}
\value{
an object of class \code{data.frame} in tidy format
}
\description{
Load in hctsa formatted MATLAB files of time series data into a tidy format ready for feature extraction
}
\author{
Trent Henderson
}
