% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{quantiles}
\alias{quantiles}
\title{Calculate a basic set of quantiles for an input time-series vector}
\usage{
quantiles(y, quantiles = seq(0, 1, by = 0.01))
}
\arguments{
\item{y}{\code{numeric} vector of values}

\item{quantiles}{\code{numeric} vector of quantiles to calculate. Defaults to \code{seq(0, 1, by = 0.01)}}
}
\value{
\code{data.frame} of results
}
\description{
Calculate a basic set of quantiles for an input time-series vector
}
\author{
Trent Henderson
}
