% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_data.R
\name{resample_data}
\alias{resample_data}
\title{Helper function to create a resampled dataset}
\usage{
resample_data(data, train_rows, test_rows, train_groups, test_groups, seed)
}
\arguments{
\item{data}{\code{data.frame} containing time-series features}

\item{train_rows}{\code{integer} denoting the number of cases in the train set}

\item{test_rows}{\code{integer} denoting the number of cases in the test set}

\item{train_groups}{\code{data.frame} containing proportions of each class in original train split}

\item{test_groups}{\code{data.frame} containing proportions of each class in original test split}

\item{seed}{\code{integer} denoting fixed value for R's pseudorandom number generator}
}
\value{
\code{list} containing new train and test data
}
\description{
Helper function to create a resampled dataset
}
\author{
Trent Henderson
}
