% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampled_ttest.R
\name{resampled_ttest}
\alias{resampled_ttest}
\title{Compute correlated t-statistic and p-value for resampled data from {correctR} package}
\usage{
resampled_ttest(x, y, n, n1, n2)
}
\arguments{
\item{x}{\code{numeric} vector of values for model A}

\item{y}{\code{numeric} vector of values for model B}

\item{n}{\code{integer} denoting number of repeat samples. Defaults to \code{length(x)}}

\item{n1}{\code{integer} denoting train set size}

\item{n2}{\code{integer} denoting test set size}
}
\value{
object of class \code{data.frame}
}
\description{
Compute correlated t-statistic and p-value for resampled data from {correctR} package
}
\references{
Nadeau, C., and Bengio, Y. Inference for the Generalization Error. Machine Learning 52, (2003).

Bouckaert, R. R., and Frank, E. Evaluating the Replicability of Significance Tests for Comparing Learning Algorithms. Advances in Knowledge Discovery and Data Mining. PAKDD 2004. Lecture Notes in Computer Science, 3056, (2004).
}
\author{
Trent Henderson
}
