% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rescale_vals.R
\name{rescale_zscore}
\alias{rescale_zscore}
\title{Calculate z-score for all columns in a dataset using train set central tendency and spread}
\usage{
rescale_zscore(data, rescalers)
}
\arguments{
\item{data}{\code{data.frame} containing data to normalise}

\item{rescalers}{\code{list} containing central tendency and spread values for the train set}
}
\value{
\code{data.frame} of rescaled data
}
\description{
Calculate z-score for all columns in a dataset using train set central tendency and spread
}
\author{
Trent Henderson
}
