% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_best_features.R
\name{select_best_features}
\alias{select_best_features}
\title{Perform a fast reduction of features based on an output vector using analysis of variance or correlation}
\usage{
select_best_features(
  data,
  k = floor(length(unique(data[[1]]$names))/2),
  outputs = NULL
)
}
\arguments{
\item{data}{the \code{feature_calculations} object containing the raw feature matrix produced by \code{calculate_features}}

\item{k}{\code{integer} denoting the number of features to retain. Defaults to half the length of the unique features available in \code{data}}

\item{outputs}{\code{data.frame} containing output data if it was not included in the \code{group} column initially when running \code{calculate_features}. Should have 2 columns where the first is an ID variable that can be joined to the \code{id} column of \code{data} and the second the output of interest (i.e., a 'y' variable). If a character or factor, classification is assumed, and if numeric, regression is assumed. Defaults to \code{NULL} which assumes the \code{group} variable exists in \code{data}}
}
\value{
object of class \code{feature_calculations} that contains the summary statistics for each feature that was retained
}
\description{
Perform a fast reduction of features based on an output vector using analysis of variance or correlation
}
\author{
Trent Henderson
}
