% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simData}
\alias{simData}
\title{Sample of randomly-generated time series to produce function tests and vignettes}
\format{
A tidy tsibble with 4 variables:
\describe{
  \item{id}{Unique identifier for the time series}
  \item{timepoint}{Time index}
  \item{values}{Value}
  \item{process}{Group label for the type of time series}
}
}
\usage{
simData
}
\description{
The variables include:
}
\keyword{datasets}
