% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_calcs.R
\name{skewness}
\alias{skewness}
\title{Calculate a skewness of a vector}
\usage{
skewness(y)
}
\arguments{
\item{y}{\code{numeric} vector of values}
}
\value{
\code{numeric} scalar of skewness
}
\description{
Calculate a skewness of a vector
}
\author{
Trent Henderson
}
