% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalers.R
\name{zscore_scaler}
\alias{zscore_scaler}
\title{Rescales a numeric vector into z-scores and then into the unit interval [0,1]}
\usage{
zscore_scaler(x, unitInt = TRUE)
}
\arguments{
\item{x}{\code{numeric vector}}

\item{unitInt}{\code{Boolean} whether to rescale into unit interval \code{[0,1]}. Defaults to \code{TRUE}}
}
\value{
\code{numeric} vector
}
\description{
\eqn{z_{i} = \frac{x_{i} - \mu}{\sigma}}
}
\author{
Trent Henderson
}
