% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_features.R
\name{calculate_features}
\alias{calculate_features}
\title{Automatically run time-series feature calculations included in the package}
\usage{
calculate_features(
  data,
  id_var = NULL,
  time_var = NULL,
  values_var = NULL,
  feature_set = c("all", "catch22", "feasts", "tsfeatures")
)
}
\arguments{
\item{data}{a dataframe with at least 4 columns: id variable, group variable, time variable, value variable}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to NULL}

\item{time_var}{a string specifying the time index variable. Defaults to NULL}

\item{feature_set}{The set of time-series features to calculate. Defaults to 'all'}
}
\value{
object of class DataFrame that contains the summary statistics for each feature
}
\description{
Automatically run time-series feature calculations included in the package
}
\examples{
\dontrun{
library(dplyr)
d <- tsibbledata::aus_retail \%>\%
  filter(State == "New South Wales")
outs <- calculate_features(data = d, id_var = "Industry", time_var = "Month", values_var = "Turnover", feature_set = "all")
}

}
\author{
Trent Henderson
}
