% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_features.R
\name{calculate_features}
\alias{calculate_features}
\title{Automatically run time-series feature calculations included in the package}
\usage{
calculate_features(
  data,
  id_var = NULL,
  group_var = NULL,
  time_var = NULL,
  value_var = NULL,
  feature_set = c("all", "catch22", "catchaMouse16", "feasts", "tsfeatures")
)
}
\arguments{
\item{data}{a dataframe with at least 4 columns: id variable, group variable, time variable, value variable}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to NULL}

\item{group_var}{a string specifying the grouping variable that the data aggregates to. Defaults to NULL}

\item{feature_set}{The set of time-series features to calculate. Defaults to 'all'}
}
\value{
object of class DataFrame that contains the summary statistics for each feature
}
\description{
Automatically run time-series feature calculations included in the package
}
\examples{
\dontrun{
library(dplyr)
d <- data.frame(value = 1 + 0.5 * 1:1000 + arima.sim(list(ma = 0.5), n = 1000)) \%>\%
dplyr::mutate(id = 1,
              group = 1,
              timepoint = row_number())
outs <- calculate_features(data = d, id_var = "id", group_var = "group", time_var = "timepoint", value_var = "value", feature_set = "all")
}

}
\author{
Trent Henderson
}
