% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_features.R
\name{calculate_features}
\alias{calculate_features}
\title{Compute features on an input time series dataset}
\usage{
calculate_features(
  data,
  id_var = NULL,
  time_var = NULL,
  values_var = NULL,
  group_var = NULL,
  feature_set = c("all", "catch22", "feasts", "tsfeatures", "tsfresh", "tsfel"),
  tsfresh_cleanup = FALSE
)
}
\arguments{
\item{data}{a dataframe with at least 4 columns: id variable, group variable, time variable, value variable}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to NULL}

\item{time_var}{a string specifying the time index variable. Defaults to NULL}

\item{values_var}{a string specifying the values variable. Defaults to NULL}

\item{group_var}{a string specifying the grouping variable that each unique series sits under. Defaults to NULL}

\item{feature_set}{the set of time-series features to calculate. Defaults to 'all'}

\item{tsfresh_cleanup}{a Boolean specifying whether to use the in-built 'tsfresh' relevant feature filter or not. Defaults to FALSE}
}
\value{
object of class DataFrame that contains the summary statistics for each feature
}
\description{
Compute features on an input time series dataset
}
\examples{
\dontrun{
featMat <- calculate_features(data = simData, 
  id_var = "id", 
  time_var = "timepoint", 
  values_var = "values", 
  group_var = "process", 
  feature_set = "catch22")
}

}
\author{
Trent Henderson
}
