% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_features.R
\name{calculate_features}
\alias{calculate_features}
\title{Automatically run time-series feature calculations included in the package}
\usage{
calculate_features(
  data,
  id_var = NULL,
  time_var = NULL,
  values_var = NULL,
  group_var = NULL,
  feature_set = c("all", "catch22", "feasts", "tsfeatures", "tsfresh", "tsfel"),
  tsfresh_cleanup = FALSE
)
}
\arguments{
\item{data}{a dataframe with at least 4 columns: id variable, group variable, time variable, value variable}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to NULL}

\item{time_var}{a string specifying the time index variable. Defaults to NULL}

\item{values_var}{a string specifying the values variable. Defaults to NULL}

\item{group_var}{a string specifying the grouping variable that the data aggregates to. Defaults to NULL}

\item{feature_set}{the set of time-series features to calculate. Defaults to 'all'}

\item{tsfresh_cleanup}{a Boolean specifying whether to use the in-built 'tsfresh' relevant feature filter or not. Defaults to FALSE}
}
\value{
object of class DataFrame that contains the summary statistics for each feature
}
\description{
Automatically run time-series feature calculations included in the package
}
\examples{
\dontrun{
library(dplyr)
library(tsibbledata)
d <- tsibbledata::aus_retail \%>\%
  filter(State == "New South Wales")
outs <- calculate_features(data = d, id_var = "Industry", time_var = "Month", 
  values_var = "Turnover", group_var = NULL, feature_set = "all", 
  tsfresh_cleanup = FALSE)
}

}
\author{
Trent Henderson
}
