% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vector_quality.R
\name{check_vector_quality}
\alias{check_vector_quality}
\title{Check data quality of a vector}
\usage{
check_vector_quality(x)
}
\arguments{
\item{x}{input data vector}
}
\value{
Boolean of whether the data is good to extract features on or not
}
\description{
Check data quality of a vector
}
\examples{
\dontrun{
x <- stats::rnorm(x)
check_vector_quality(x)
}

}
