% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_top_features.R
\name{compute_top_features}
\alias{compute_top_features}
\title{Return an object containing results from top-performing features on a classification task}
\usage{
compute_top_features(
  data,
  id_var = "id",
  group_var = "group",
  num_features = 40,
  normalise_violin_plots = FALSE,
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax"),
  cor_method = c("pearson", "spearman"),
  test_method = "gaussprRadial",
  use_empirical_null = FALSE,
  use_k_fold = FALSE,
  num_folds = 0,
  split_prop = 0.8,
  num_shuffles = 50,
  pool_empirical_null = FALSE
)
}
\arguments{
\item{data}{the dataframe containing the raw feature matrix}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to "id"}

\item{group_var}{a string specifying the grouping variable that the data aggregates to. Defaults to "group"}

\item{num_features}{the number of top features to retain and explore. Defaults to 40}

\item{normalise_violin_plots}{a Boolean of whether to normalise features before plotting. Defaults to FALSE}

\item{method}{a rescaling/normalising method to apply. Defaults to 'RobustSigmoid'}

\item{cor_method}{the correlation method to use. Defaults to 'pearson'}

\item{test_method}{the algorithm to use for quantifying class separation}

\item{use_empirical_null}{a Boolean specifying whether to use empirical null procedures to compute p-values if linear svm or rbf svm is selected. Defaults to FALSE}

\item{use_k_fold}{a Boolean specifying whether to use k-fold procedures for generating a distribution of classification accuracy estimates. Defaults to FALSE}

\item{num_folds}{an integer specifying the number of k-folds to perform if data has more than two classes.Defaults to 10}

\item{split_prop}{a double between 0 and 1 specifying the proportion of input data that should go into the training set (therefore 1 - p goes into the test set). Defaults to 0.8}

\item{num_shuffles}{an integer specifying the number of class label shuffles to perform if linear svm or rbf svm is selected. Defaults to 5}

\item{pool_empirical_null}{a Boolean specifying whether to use the pooled empirical null distribution of all features or each features' individual empirical null distribution if linear svm or rbf svm is selected and use_empirical_null is TRUE. Defaults to FALSE}
}
\value{
an object of class list containing a dataframe of results, a feature x feature matrix plot, and a violin plot
}
\description{
Return an object containing results from top-performing features on a classification task
}
\examples{
\dontrun{
featMat <- calculate_features(data = simData, 
  id_var = "id", 
  time_var = "timepoint", 
  values_var = "values", 
  group_var = "process", 
  feature_set = "catch22")
  
compute_top_features(featMat,
  id_var = "id",
  group_var = "group",
  num_features = 10,
  normalise_violin_plots = FALSE,
  method = "RobustSigmoid",
  cor_method = "pearson",
  test_method = "linear svm",
  use_empirical_null = FALSE,
  use_k_fold = FALSE,
  num_folds = 10,
  split_prop = 0.8,
  num_shuffles = 50,
  pool_empirical_null = FALSE) 
}

}
\author{
Trent Henderson
}
