% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_duplicates.R
\name{filter_duplicates}
\alias{filter_duplicates}
\title{Remove duplicate features that exist in multiple feature sets and retain a reproducible random selection of one of them}
\usage{
filter_duplicates(data, preference = NULL, seed = 123)
}
\arguments{
\item{data}{\code{feature_calculations} object containing the raw feature matrix produced by \code{calculate_features}}

\item{preference}{deprecated. Do not use}

\item{seed}{\code{integer} denoting a fix for R's pseudo-random number generator to ensure selections are reproducible. Defaults to \code{123}}
}
\value{
\code{feature_calculations} object containing filtered feature data
}
\description{
Remove duplicate features that exist in multiple feature sets and retain a reproducible random selection of one of them
}
\author{
Trent Henderson
}
