% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_feature_classifier.R
\name{fit_feature_classifier}
\alias{fit_feature_classifier}
\title{Fit a classifier to feature matrix to extract top performers}
\usage{
fit_feature_classifier(
  data,
  id_var = "id",
  group_var = "group",
  test_method = c("t-test", "binomial logistic", "linear svm", "rbf svm")
)
}
\arguments{
\item{data}{the dataframe containing the raw feature matrix}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to "id"}

\item{group_var}{a string specifying the grouping variable that the data aggregates to. Defaults to "group"}

\item{test_method}{the algorithm to use for quantifying class separation}
}
\value{
an object of class dataframe containing results
}
\description{
Fit a classifier to feature matrix to extract top performers
}
\examples{
\dontrun{
featMat <- calculate_features(data = simData, 
  id_var = "id", 
  time_var = "timepoint", 
  values_var = "values", 
  group_var = "process", 
  feature_set = "catch22")
  
fit_feature_classifier(featMat,
  id_var = "id",
  group_var = "group") 
}

}
\author{
Trent Henderson
}
