% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_feature_classifier.R
\name{fit_feature_classifier}
\alias{fit_feature_classifier}
\title{Fit a classifier to feature matrix to extract top performers}
\usage{
fit_feature_classifier(
  data,
  id_var = "id",
  group_var = "group",
  test_method = c("t-test", "wilcox", "binomial logistic", "linear svm", "rbf svm"),
  use_empirical_null = FALSE,
  use_k_fold = FALSE,
  num_folds = 0,
  num_shuffles = 50,
  pool_empirical_null = FALSE
)
}
\arguments{
\item{data}{the dataframe containing the raw feature matrix}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to "id"}

\item{group_var}{a string specifying the grouping variable that the data aggregates to. Defaults to "group"}

\item{test_method}{the algorithm to use for quantifying class separation}

\item{use_empirical_null}{a Boolean specifying whether to use empirical null procedures to compute p-values if linear svm or rbf svm is selected. Defaults to FALSE}

\item{use_k_fold}{a Boolean specifying whether to use k-fold procedures for generating a distribution of classification accuracy estimates. Defaults to FALSE}

\item{num_folds}{an integer specifying the number of folds (train-test splits) to perform if linear svm or rbf svm is selected and use_k_fold is set to TRUE. Defaults to 0 but if TRUE, 10 is recommended}

\item{num_shuffles}{an integer specifying the number of class label shuffles to perform if linear svm or rbf svm is selected. Defaults to 50}

\item{pool_empirical_null}{a Boolean specifying whether to use the pooled empirical null distribution of all features or each features' individual empirical null distribution if linear svm or rbf svm is selected and use_empirical_null is TRUE. Defaults to FALSE}
}
\value{
an object of class dataframe containing results
}
\description{
Fit a classifier to feature matrix to extract top performers
}
\examples{
\dontrun{
featMat <- calculate_features(data = simData, 
  id_var = "id", 
  time_var = "timepoint", 
  values_var = "values", 
  group_var = "process", 
  feature_set = "catch22")
  
fit_feature_classifier(featMat,
  id_var = "id",
  group_var = "group",
  test_method = "linear svm",
  use_empirical_null = TRUE,
  use_k_fold = TRUE,
  num_folds = 10,
  num_shuffles = 5,
  pool_empirical_null = FALSE) 
}

}
\author{
Trent Henderson
}
