% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_multivariate_classifier.R
\name{fit_multivariate_classifier}
\alias{fit_multivariate_classifier}
\title{Fit a classifier to feature matrix using all features or all features by set}
\usage{
fit_multivariate_classifier(
  data,
  id_var = "id",
  group_var = "group",
  by_set = FALSE,
  test_method = c("linear svm", "rbf svm"),
  use_empirical_null = FALSE,
  use_k_fold = FALSE,
  num_folds = 0,
  num_shuffles = 50
)
}
\arguments{
\item{data}{the dataframe containing the raw feature matrix}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to "id"}

\item{group_var}{a string specifying the grouping variable that the data aggregates to. Defaults to "group"}

\item{by_set}{Boolean specifying whether to compute classifiers for each feature set. Defaults to FALSE}

\item{test_method}{the algorithm to use for quantifying class separation}

\item{use_empirical_null}{a Boolean specifying whether to use empirical null procedures to compute p-values if linear svm or rbf svm is selected. Defaults to FALSE}

\item{use_k_fold}{a Boolean specifying whether to use k-fold procedures for generating a distribution of classification accuracy estimates. Defaults to FALSE}

\item{num_folds}{an integer specifying the number of folds (train-test splits) to perform if use_k_fold is set to TRUE. Defaults to 10}

\item{num_shuffles}{an integer specifying the number of class label shuffles to perform. Defaults to 50}
}
\value{
an object of class list containing summaries of the classification models
}
\description{
Fit a classifier to feature matrix using all features or all features by set
}
\examples{
\dontrun{
featMat <- calculate_features(data = simData, 
  id_var = "id", 
  time_var = "timepoint", 
  values_var = "values", 
  group_var = "process", 
  feature_set = "catch22")
  
fit_multivariate_classifier(featMat,
  id_var = "id",
  group_var = "group",
  by_set = FALSE,
  test_method = "linear svm",
  use_empirical_null = TRUE,
  use_k_fold = TRUE,
  num_folds = 10,
  num_shuffles = 50) 
}

}
\author{
Trent Henderson
}
