% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_theft.R
\name{init_theft}
\alias{init_theft}
\title{Wrapper function to communicate to R the Python virtual environment containing the relevant libraries for calculating features}
\usage{
init_theft(python_path, venv_path)
}
\arguments{
\item{python_path}{\code{character} specifying the filepath to the version of Python you wish to use}

\item{venv_path}{\code{character} specifying the filepath to the Python virtual environment where "tsfresh", "tsfel", and/or "kats" are installed}
}
\description{
Wrapper function to communicate to R the Python virtual environment containing the relevant libraries for calculating features
}
\author{
Trent Henderson
}
