% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_theft.R
\name{init_theft}
\alias{init_theft}
\title{Communicate to R the Python virtual environment containing the relevant libraries for calculating features}
\usage{
init_theft(venv)
}
\arguments{
\item{venv}{\code{character} specifying the name of the to the Python virtual environment where \code{"tsfresh"}, \code{"TSFEL"}, and/or \code{"Kats"} are installed}
}
\value{
no return value; called for side effects
}
\description{
Communicate to R the Python virtual environment containing the relevant libraries for calculating features
}
\author{
Trent Henderson
}
