% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_pkgs.R
\name{install_python_pkgs}
\alias{install_python_pkgs}
\title{Download and install all the relevant Python packages into a target location}
\usage{
install_python_pkgs(venv, simple_kats = TRUE)
}
\arguments{
\item{venv}{\code{character} specifying the name of the new virtual environment to create}

\item{simple_kats}{\code{Boolean} denoting whether to try a standard installation of \code{Kats} from PyPI using \code{reticulate::virtualenv_install} or to install a safer version with less dependencies. Defaults to \code{TRUE}}
}
\description{
Download and install all the relevant Python packages into a target location
}
\author{
Trent Henderson
}
