% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalers.R
\name{minmax_scaler}
\alias{minmax_scaler}
\title{This function rescales a vector of numerical values into the unit interval [0,1]}
\usage{
minmax_scaler(x)
}
\arguments{
\item{x}{a numeric vector, preferably of feature values computed by other {theft} package functions}
}
\value{
x a numeric vector, rescaled into the [0,1] unit interval
}
\description{
This function rescales a vector of numerical values into the unit interval [0,1]
}
\examples{
\dontrun{
x <- 1 + 0.5 * 1:1000 + arima.sim(list(ma = 0.5), n = 1000)
outs <- minmax_scaler(x)
}

}
\author{
Trent Henderson
}
