% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalers.R
\name{minmax_scaler}
\alias{minmax_scaler}
\title{Rescales a numeric vector into the unit interval [0,1]}
\usage{
minmax_scaler(x)
}
\arguments{
\item{x}{\code{numeric} vector}
}
\value{
\code{numeric} vector
}
\description{
\eqn{z_{i} = \frac{x_{i} - \text{min}(\mathbf{x})}{\text{max}(\mathbf{x}) - \text{min}(\mathbf{x})}}
}
\author{
Trent Henderson
}
