% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Scale each feature vector into a user-specified range for visualisation and modelling}
\usage{
normalise(data, method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax"))
}
\arguments{
\item{data}{either a \code{feature_calculations} object containing the raw feature matrix produced by \code{calculate_features} or a \code{vector} of class \code{numeric} containing values to be normalised}

\item{method}{a rescaling/normalising method to apply to violin plots. Defaults to \code{"z-score"}}
}
\value{
either an object of class \code{data.frame} or \code{numeric}
}
\description{
Scale each feature vector into a user-specified range for visualisation and modelling
}
\author{
Trent Henderson
}
