% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.feature_calculations}
\alias{plot.feature_calculations}
\title{Produce a plot for a feature_calculations object}
\usage{
\method{plot}{feature_calculations}(
  data,
  type = c("quality", "matrix", "cor"),
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax"),
  clust_method = c("average", "ward.D", "ward.D2", "single", "complete", "mcquitty",
    "median", "centroid")
)
}
\arguments{
\item{data}{the \code{feature_calculations} object containing the raw feature matrix produced by \code{calculate_features}}

\item{type}{\code{string} specifying the type of plot to draw. Defaults to \code{"quality"}}

\item{method}{a rescaling/normalising method to apply. Defaults to \code{"z-score"}}

\item{clust_method}{the hierarchical clustering method to use for the pairwise correlation plot. Defaults to \code{"average"}}
}
\value{
an object of class \code{ggplot} that contains the heatmap graphic
}
\description{
Produce a plot for a feature_calculations object
}
\author{
Trent Henderson
}
