% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.low_dimension}
\alias{plot.low_dimension}
\title{Produce a plot for a low_dimension object}
\usage{
\method{plot}{low_dimension}(data, show_covariance = TRUE, ...)
}
\arguments{
\item{data}{the \code{low_dimension} object containing the raw feature matrix produced by \code{reduce_dims}}

\item{show_covariance}{a Boolean as to whether covariance ellipses should be shown on the plot. Defaults to \code{FALSE}}
}
\value{
an object of class \code{ggplot} that contains the heatmap graphic
}
\description{
Produce a plot for a low_dimension object
}
\author{
Trent Henderson
}
