% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_connectivity_matrix.R
\name{plot_connectivity_matrix}
\alias{plot_connectivity_matrix}
\title{Produce a correlation matrix plot showing pairwise correlations of feature vectors by unique id with automatic hierarchical clustering.
This function is based on functional connectivity matrices in neuroscience.}
\usage{
plot_connectivity_matrix(
  data,
  id_var = NULL,
  names_var = NULL,
  values_var = NULL
)
}
\arguments{
\item{data}{a dataframe with at least 3 columns for 'id', 'names' and 'values'}

\item{id_var}{a string specifying the ID variable to compute pairwise correlations between. Defaults to NULL}

\item{names_var}{a string denoting the name of the variable/column that holds the feature names}

\item{values_var}{a string denoting the name of the variable/column that holds the numerical feature values}
}
\value{
an object of class ggplot that contains the correlation matrix graphic
}
\description{
Produce a correlation matrix plot showing pairwise correlations of feature vectors by unique id with automatic hierarchical clustering.
This function is based on functional connectivity matrices in neuroscience.
}
\examples{
\dontrun{
library(dplyr)
library(tsibbledata)
d <- tsibbledata::aus_retail \%>\%
  filter(State == "New South Wales")
outs <- calculate_features(data = d, id_var = "Industry", time_var = "Month", 
  values_var = "Turnover", feature_set = "all", tsfresh_cleanup = FALSE)
normed <- normalise_feature_frame(outs, names_var = "names", 
  values_var = "values", method = "RobustSigmoid")
plot_connectivity_matrix(normed, id_var = "Industry")
}

}
\author{
Trent Henderson
}
