% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_correlations.R
\name{plot_feature_correlations}
\alias{plot_feature_correlations}
\title{Produce a correlation matrix plot showing pairwise correlations of feature vectors by unique id with automatic hierarchical clustering.}
\usage{
plot_feature_correlations(
  data,
  is_normalised = FALSE,
  id_var = "id",
  names_var = "names",
  values_var = "values",
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax"),
  cor_method = c("pearson", "spearman"),
  interactive = FALSE
)
}
\arguments{
\item{data}{a dataframe with at least 3 columns for \code{'id'}, \code{'names'} and \code{'values'}}

\item{is_normalised}{a Boolean as to whether the input feature values have already been scaled. Defaults to \code{FALSE}}

\item{id_var}{a string specifying the ID variable to compute pairwise correlations between. Defaults to \code{"id"}}

\item{names_var}{a string denoting the name of the variable/column that holds the feature names. Defaults to \code{"names"}}

\item{values_var}{a string denoting the name of the variable/column that holds the numerical feature values. Defaults to \code{"values"}}

\item{method}{a rescaling/normalising method to apply. Defaults to \code{"RobustSigmoid"}}

\item{cor_method}{the correlation method to use. Defaults to \code{"pearson"}}

\item{interactive}{a Boolean as to whether to plot an interactive \code{plotly} graphic. Defaults to \code{FALSE}}
}
\value{
an object of class \code{ggplot} that contains the correlation matrix graphic
}
\description{
Produce a correlation matrix plot showing pairwise correlations of feature vectors by unique id with automatic hierarchical clustering.
}
\examples{
featMat <- calculate_features(data = simData, 
  id_var = "id", 
  time_var = "timepoint", 
  values_var = "values", 
  group_var = "process", 
  feature_set = "catch22")
  
plot_feature_correlations(data = featMat, 
  is_normalised = FALSE, 
  id_var = "id", 
  names_var = "names", 
  values_var = "values",
  method = "RobustSigmoid",
  interactive = FALSE)

}
\author{
Trent Henderson
}
