% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_matrix.R
\name{plot_feature_matrix}
\alias{plot_feature_matrix}
\title{Produce a heatmap matrix of the calculated feature value vectors and each unique time series with automatic hierarchical clustering.}
\usage{
plot_feature_matrix(
  data,
  is_normalised = FALSE,
  id_var = NULL,
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax", "MeanSubtract")
)
}
\arguments{
\item{data}{a dataframe with at least 2 columns called 'names' and 'values'}

\item{is_normalised}{a Boolean as to whether the input feature values have already been scaled. Defaults to FALSE}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to NULL}

\item{method}{a rescaling/normalising method to apply. Defaults to 'RobustSigmoid'}
}
\value{
an object of class ggplot that contains the heatmap graphic
}
\description{
Produce a heatmap matrix of the calculated feature value vectors and each unique time series with automatic hierarchical clustering.
}
\examples{
\dontrun{
data1 <- 1 + 0.5 * 1:1000 + arima.sim(list(ma = 0.5), n = 1000)
data2 <- rnorm(1000, mean = 0, sd = 1)
outs1 <- catch22_all(data1)
outs1['group'] <- 'Group 1'
outs2 <- catch22_all(data2)
outs2['group'] <- 'Group 2'
outs <- rbind(outs1, outs2)
plot_feature_matrix(outs, is_normalised = FALSE, id_var = "group", method = "RobustSigmoid")
}

}
\seealso{
[catchEmAll::normalise_catch()]
}
\author{
Trent Henderson
}
