% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_matrix.R
\name{plot_feature_matrix}
\alias{plot_feature_matrix}
\title{Produce a heatmap matrix of the calculated feature value vectors and each unique time series with automatic hierarchical clustering.}
\usage{
plot_feature_matrix(
  data,
  is_normalised = NULL,
  id_var = "id",
  method = NULL,
  clust_method = c("average", "ward.D", "ward.D2", "single", "complete", "mcquitty",
    "median", "centroid"),
  interactive = FALSE
)
}
\arguments{
\item{data}{a dataframe with at least 2 columns called \code{"names"} and \code{"values"}}

\item{is_normalised}{deprecated as of 0.4.0; do not use}

\item{id_var}{a string specifying the ID variable to identify each time series. Defaults to \code{"id"}}

\item{method}{deprecated as of 0.4.0; do not use}

\item{clust_method}{the hierarchical clustering method to use for the pairwise correlation plot. Defaults to \code{"average"}}

\item{interactive}{a Boolean as to whether to plot an interactive \code{plotly} graphic. Defaults to \code{FALSE}}
}
\value{
an object of class \code{ggplot} that contains the heatmap graphic
}
\description{
Produce a heatmap matrix of the calculated feature value vectors and each unique time series with automatic hierarchical clustering.
}
\examples{
featMat <- calculate_features(data = simData, 
  id_var = "id", 
  time_var = "timepoint", 
  values_var = "values", 
  group_var = "process", 
  feature_set = "catch22",
  seed = 123)

plot_feature_matrix(featMat, 
  id_var = "id", 
  clust_method = "average",
  interactive = FALSE)

}
\author{
Trent Henderson
}
