% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_matrix.R
\name{plot_feature_matrix}
\alias{plot_feature_matrix}
\title{Produce a heatmap matrix of the calculated feature value vectors and each unique time series with automatic hierarchical clustering.}
\usage{
plot_feature_matrix(
  data,
  is_normalised = FALSE,
  id_var = NULL,
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax", "MeanSubtract"),
  interactive = FALSE
)
}
\arguments{
\item{data}{a dataframe with at least 2 columns called 'names' and 'values'}

\item{is_normalised}{a Boolean as to whether the input feature values have already been scaled. Defaults to FALSE}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to NULL}

\item{method}{a rescaling/normalising method to apply. Defaults to 'RobustSigmoid'}

\item{interactive}{a Boolean as to whether to plot an interactive plotly graphic. Defaults to FALSE}
}
\value{
an object of class ggplot that contains the heatmap graphic
}
\description{
Produce a heatmap matrix of the calculated feature value vectors and each unique time series with automatic hierarchical clustering.
}
\examples{
\dontrun{
library(dplyr)
library(tsibbledata)

d <- tsibbledata::aus_retail \%>\%
  rename(Series_ID = 3)

feature_matrix <- calculate_features(data = d, 
  id_var = "Series_ID", 
  time_var = "Month", 
  values_var = "Turnover", 
  group_var = "State",
  feature_set = "catch22")
}

plot_feature_matrix(feature_matrix, is_normalised = FALSE, 
  id_var = "id", 
  method = "MinMax",
  interactive = FALSE)

}
\author{
Trent Henderson
}
