% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_low_dimension.R
\name{plot_low_dimension}
\alias{plot_low_dimension}
\title{Produce a principal components analysis (PCA) on normalised feature values and render a bivariate plot to visualise it}
\usage{
plot_low_dimension(
  data,
  is_normalised = FALSE,
  id_var = NULL,
  group_var = NULL,
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax", "MeanSubtract"),
  low_dim_method = c("PCA", "t-SNE"),
  perplexity = 30,
  plot = TRUE
)
}
\arguments{
\item{data}{a dataframe with at least 2 columns called 'names' and 'values'}

\item{is_normalised}{a Boolean as to whether the input feature values have already been scaled. Defaults to FALSE}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to NULL}

\item{group_var}{a string specifying the grouping variable that the data aggregates to. Defaults to NULL}

\item{method}{a rescaling/normalising method to apply. Defaults to 'RobustSigmoid'}

\item{low_dim_method}{the low dimensional embedding method to use. Defaults to 'PCA'}

\item{perplexity}{the perplexity hyperparameter to use if t-SNE algorithm is selected. Defaults to 30}

\item{plot}{a Boolean as to whether a bivariate plot should be returned or the calculation dataframe. Defaults to TRUE}
}
\value{
if plot = TRUE, returns an object of class ggplot, if plot = FALSE returns an object of class dataframe with PCA results
}
\description{
Produce a principal components analysis (PCA) on normalised feature values and render a bivariate plot to visualise it
}
\examples{
\dontrun{
library(dplyr)
library(tsibbledata)

d <- tsibbledata::aus_retail \%>\%
  ename(Series_ID = 3)

feature_matrix <- calculate_features(data = d, 
  id_var = "Series_ID", 
  time_var = "Month", 
  values_var = "Turnover", 
  group_var = "State",
  feature_set = "catch22")
}

plot_low_dimension(feature_matrix, is_normalised = FALSE, 
 id_var = "id", 
 group_var = "State", 
 method = "MinMax", 
 low_dim_method = "PCA", 
 plot = TRUE)

}
\author{
Trent Henderson
}
