% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_low_dimension.R
\name{plot_low_dimension}
\alias{plot_low_dimension}
\title{Produce a principal components analysis (PCA) on normalised feature values and render a bivariate plot to visualise it}
\usage{
plot_low_dimension(
  data,
  is_normalised = FALSE,
  id_var = NULL,
  group_var = NULL,
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax", "MeanSubtract"),
  plot = TRUE
)
}
\arguments{
\item{data}{a dataframe with at least 2 columns called 'names' and 'values'}

\item{is_normalised}{a Boolean as to whether the input feature values have already been scaled. Defaults to FALSE}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to NULL}

\item{group_var}{a string specifying the grouping variable that the data aggregates to. Defaults to NULL}

\item{method}{a rescaling/normalising method to apply. Defaults to 'RobustSigmoid'}

\item{plot}{a Boolean as to whether a bivariate plot should be returned or the calculation dataframe. Defaults to TRUE}
}
\value{
if plot = TRUE, returns an object of class ggplot, if plot = FALSE returns an object of class dataframe with PCA results
}
\description{
Produce a principal components analysis (PCA) on normalised feature values and render a bivariate plot to visualise it
}
\examples{
\dontrun{
library(dplyr)
d <- tsibbledata::aus_retail \%>\%
  filter(State == "New South Wales")
outs <- calculate_features(data = d, id_var = "Industry", time_var = "Month", values_var = "Turnover", feature_set = "all")
plot_low_dimension(outs, is_normalised = FALSE, id_var = "Industry", group_var = NULL, method = "RobustSigmoid")
}

}
\author{
Trent Henderson
}
