% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_low_dimension.R
\name{plot_low_dimension}
\alias{plot_low_dimension}
\title{Produce a principal components analysis (PCA) on normalised feature values and render a bivariate plot to visualise it}
\usage{
plot_low_dimension(
  data,
  is_normalised = FALSE,
  id_var = "id",
  group_var = "group",
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax", "MeanSubtract"),
  low_dim_method = c("PCA", "t-SNE"),
  perplexity = 30,
  plot = TRUE,
  show_covariance = FALSE
)
}
\arguments{
\item{data}{a dataframe with at least 2 columns called 'names' and 'values'}

\item{is_normalised}{a Boolean as to whether the input feature values have already been scaled. Defaults to FALSE}

\item{id_var}{a string specifying the ID variable to group data on (if one exists). Defaults to "id"}

\item{group_var}{a string specifying the grouping variable that the data aggregates to. Defaults to "group"}

\item{method}{a rescaling/normalising method to apply. Defaults to 'RobustSigmoid'}

\item{low_dim_method}{the low dimensional embedding method to use. Defaults to 'PCA'}

\item{perplexity}{the perplexity hyperparameter to use if t-SNE algorithm is selected. Defaults to 30}

\item{plot}{a Boolean as to whether a bivariate plot should be returned or the calculation dataframe. Defaults to TRUE}

\item{show_covariance}{a Boolean as to whether covariance ellipses should be shown on the plot. Defaults to FALSE}
}
\value{
if plot = TRUE, returns an object of class ggplot, if plot = FALSE returns an object of class dataframe with PCA results
}
\description{
Produce a principal components analysis (PCA) on normalised feature values and render a bivariate plot to visualise it
}
\examples{
\dontrun{
featMat <- calculate_features(data = simData, 
  id_var = "id", 
  time_var = "timepoint", 
  values_var = "values", 
  group_var = "process", 
  feature_set = "catch22")

plot_low_dimension(featMat, 
  is_normalised = FALSE, 
  id_var = "id", 
  group_var = "State", 
  method = "RobustSigmoid", 
  low_dim_method = "PCA", 
  plot = TRUE,
  show_covariance = TRUE)
}

}
\author{
Trent Henderson
}
