% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quality_matrix.R
\name{plot_quality_matrix}
\alias{plot_quality_matrix}
\title{Produce a matrix visualisation of data types computed by feature calculation function.}
\usage{
plot_quality_matrix(data)
}
\arguments{
\item{data}{a dataframe with at least 2 columns called 'names' and 'values' such as the output of catch22::catch22_all}
}
\value{
an object of class ggplot that contains the graphic
}
\description{
Produce a matrix visualisation of data types computed by feature calculation function.
}
\examples{
\dontrun{
library(dplyr)
library(tsibbledata)

d <- tsibbledata::aus_retail \%>\%
  ename(Series_ID = 3)

feature_matrix <- calculate_features(data = d, 
  id_var = "Series_ID", 
  time_var = "Month", 
  values_var = "Turnover", 
  group_var = "State",
  feature_set = "catch22")
}

plot_quality_matrix(data = feature_matrix)

}
\author{
Trent Henderson
}
