% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quality_matrix.R
\name{plot_quality_matrix}
\alias{plot_quality_matrix}
\title{Produce a matrix visualisation of data types computed by feature calculation function.}
\usage{
plot_quality_matrix(data)
}
\arguments{
\item{data}{a dataframe with at least 2 columns called 'names' and 'values' such as the output of catch22::catch22_all}
}
\value{
an object of class ggplot that contains the graphic
}
\description{
Produce a matrix visualisation of data types computed by feature calculation function.
}
\examples{
\dontrun{
library(dplyr)
d <- tsibbledata::aus_retail \%>\%
  filter(State == "New South Wales")
outs <- calculate_features(data = d, id_var = "Industry", time_var = "Month", value_var = "Turnover", feature_set = "all")
plot_quality_matrix(data = outs)
}

}
\author{
Trent Henderson
}
