% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{robustsigmoid_scaler}
\alias{robustsigmoid_scaler}
\title{This function rescales a vector of numerical values with an outlier-robust
Sigmoidal transformation using a C++ implementation for efficiency.}
\usage{
robustsigmoid_scaler(x)
}
\arguments{
\item{x}{a numeric vector, preferably of feature values computed by other functions in sawlog}
}
\value{
x a numeric vector, rescaled into Sigmoidal range
}
\description{
This function rescales a vector of numerical values with an outlier-robust
Sigmoidal transformation using a C++ implementation for efficiency.
}
\examples{
x <- 1 + 0.5 * 1:1000 + arima.sim(list(ma = 0.5), n = 1000)
outs <- robustsigmoid_scaler(x)

}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\author{
Trent Henderson
}
