% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalers.R
\name{robustsigmoid_scaler}
\alias{robustsigmoid_scaler}
\title{This function rescales a vector of numerical values with an outlier-robust Sigmoidal transformation and then into the unit interval [0,1]
\deqn{z_i = \left[1 + \exp\left(-\frac{x_i - \mathrm{median}(\mathbf{x})}{\mathrm{IQR}(\mathbf{x})/{1.35}}\right)\right]^{-1}}}
\usage{
robustsigmoid_scaler(x, unitInt = TRUE)
}
\arguments{
\item{x}{a numeric vector, preferably of feature values computed by other \code{theft} package functions}

\item{unitInt}{Booelan whether to rescale Sigmoidal outputs into unit interval \code{[0,1]}. Defaults to \code{TRUE}}
}
\value{
\code{numeric} vector
}
\description{
This function rescales a vector of numerical values with an outlier-robust Sigmoidal transformation and then into the unit interval [0,1]
\deqn{z_i = \left[1 + \exp\left(-\frac{x_i - \mathrm{median}(\mathbf{x})}{\mathrm{IQR}(\mathbf{x})/{1.35}}\right)\right]^{-1}}
}
\author{
Trent Henderson
}
