% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalers.R
\name{robustsigmoid_scaler}
\alias{robustsigmoid_scaler}
\title{Rescales a numeric vector using an outlier-robust Sigmoidal transformation}
\usage{
robustsigmoid_scaler(x)
}
\arguments{
\item{x}{\code{numeric} vector}
}
\value{
\code{numeric} vector
}
\description{
\eqn{z_{i} = \left[1 + \exp\left(-\frac{x_{i} - \text{median}(\mathbf{x})}{\text{IQR}(\mathbf{x})/{1.35}}\right)\right]^{-1}}
}
\references{
Fulcher, Ben D., Little, Max A., and Jones, Nick S. Highly Comparative Time-Series Analysis: The Empirical Structure of Time Series and Their Methods. Journal of The Royal Society Interface 10(83), (2013).
}
\author{
Trent Henderson
}
