% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalers.R
\name{sigmoid_scaler}
\alias{sigmoid_scaler}
\title{This function rescales a vector of numerical values with a Sigmoidal transformation}
\usage{
sigmoid_scaler(x, unitInt = TRUE)
}
\arguments{
\item{x}{a numeric vector, preferably of feature values computed by other \code{theft} package functions}

\item{unitInt}{Booelan whether to rescale Sigmoidal outputs into unit interval \code{[0,1]}. Defaults to \code{TRUE}}
}
\value{
x a numeric rescaled vector
}
\description{
This function rescales a vector of numerical values with a Sigmoidal transformation
}
\examples{
\dontrun{
x <- 1 + 0.5 * 1:1000 + arima.sim(list(ma = 0.5), n = 1000)
outs <- sigmoid_scaler(x)
}

}
\author{
Trent Henderson
}
