% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sigmoid_scaler}
\alias{sigmoid_scaler}
\title{This function rescales a vector of numerical values with a Sigmoidal transformation
using a C++ implementation for efficiency.}
\usage{
sigmoid_scaler(x)
}
\arguments{
\item{x}{a numeric vector, preferably of feature values computed by other catchEmAll package functions}
}
\value{
x a numeric vector, rescaled into Sigmoidal range
}
\description{
This function rescales a vector of numerical values with a Sigmoidal transformation
using a C++ implementation for efficiency.
}
\examples{
x <- 1 + 0.5 * 1:1000 + arima.sim(list(ma = 0.5), n = 1000)
outs <- sigmoid_scaler(x)

}
\author{
Trent Henderson
}
