% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalers.R
\name{sigmoid_scaler}
\alias{sigmoid_scaler}
\title{Rescales a numeric vector using a Sigmoidal transformation}
\usage{
sigmoid_scaler(x)
}
\arguments{
\item{x}{\code{numeric} vector}
}
\value{
\code{numeric} vector
}
\description{
\eqn{z_{i} = \left[1 + \exp(-\frac{x_{i} - \mu}{\sigma})\right]^{-1}}
}
\author{
Trent Henderson
}
