% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalers.R
\name{zscore_scaler}
\alias{zscore_scaler}
\title{This function rescales a vector of numerical values into z-scores}
\usage{
zscore_scaler(x, unitInt = TRUE)
}
\arguments{
\item{x}{a numeric vector, preferably of feature values computed by other \code{theft} package functions}

\item{unitInt}{Booelan whether to rescale outputs into unit interval \code{[0,1]}. Defaults to \code{TRUE}}
}
\value{
x a numeric vector, rescaled into z-scores
}
\description{
This function rescales a vector of numerical values into z-scores
}
\examples{
zscore_scaler(stats::rnorm(10))

}
\author{
Trent Henderson
}
