% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalers.R
\name{zscore_scaler}
\alias{zscore_scaler}
\title{This function rescales a vector of numerical values into z-scores}
\usage{
zscore_scaler(x)
}
\arguments{
\item{x}{a numeric vector, preferably of feature values computed by other {theft} package functions}
}
\value{
x a numeric vector, rescaled into z-scores
}
\description{
This function rescales a vector of numerical values into z-scores
}
\examples{
\dontrun{
x <- 1 + 0.5 * 1:1000 + arima.sim(list(ma = 0.5), n = 1000)
outs <- zscore_scaler(x)
}

}
\author{
Trent Henderson
}
